[Sema] Produce `OpaqueTypeArchetypeType`s in structural positions
Modify type resolution so that "some" types can be placed in structual
positions, e.g. `(some P, Q)`. This commit only allows for one structual "some"
type, but that will be loosened in a future commit. Currently type checking
will fail if the "some" type is not top-level, which will also need a fix.