[TypeChecker] Properly constrain structural `OpaqueTypeArchetypeType`s
Opaque types are now opened in `addContextualConversionConstraint` so that
constraining the type of the return expression to the return type via
`addConstraint` does the right thing. `ConstraintKind::OpaqueUnderlyingType` is
probably now no longer needed and can be removed in a future commit. Note:
opaque types in structual positions will still produce errors because the right
`UnderlyingToOpaqueExpr` conversions are not yet applied.